Separated the modules. Easier to work with, and find definitions
