+++ UI.hpp
@@ -1,0 +1,14 @@
+#include <SFML/Graphics.hpp>
+#include <SFML/Window.hpp>
+#include <iostream>
+#include "macros.hpp"
+
+using namespace sf;
+using namespace std;
+
+
+void loadBoard(RenderWindow &window);
+
+void Meniusetup(RenderWindow &meniu);
+
+void centerscreen(RenderWindow &meniu);
