+++ macros.hpp
@@ -1,0 +1,4 @@
+#define circleRadius 20.f
+#define colDist 100.f
+#define rowDist 100.f
+#define boardSize 8
